+++ projects/dsp-ui/src/lib/action/index.ts
@@ -18,6 +18,7 @@ export * from './components/message/message.component';
 export * from './components/sort-button/sort-button.component';
 export * from './components/login-form/login-form.component';
 export * from './components/string-literal-input/string-literal-input.component';
+export * from './components/confirmation-dialog/confirmation-dialog.component';
 
 // directives
 export * from './directives/admin-image/admin-image.directive';