+++ projects/dsp-ui/src/lib/viewer/views/property-view/property-view.component.spec.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { By } from '@angular/platform-browser';
 import {
     ApiResponseError,
@@ -24,7 +25,8 @@ import { PropertyViewComponent } from './property-view.component';
     <dsp-property-view #propView
       [parentResource]="parentResource"
       [propArray]="propArray"
-      [systemPropArray]="systemPropArray">
+      [systemPropArray]="systemPropArray"
+      [showAllProps]="showAllProps">
     </dsp-property-view>`
 })
 class TestPropertyParentComponent implements OnInit, OnDestroy {
@@ -36,6 +38,8 @@ class TestPropertyParentComponent implements OnInit, OnDestroy {
     propArray: PropertyInfoValues[] = [];
 
     systemPropArray: SystemPropertyDefinition[] = [];
+
+    showAllProps = false;
 
     voeSubscription: Subscription;
 
@@ -117,6 +121,7 @@ describe('PropertyViewComponent', () => {
         TestBed.configureTestingModule({
         imports: [
             MatIconModule,
+            MatTooltipModule
         ],
         declarations: [
             TestPropertyParentComponent,
@@ -214,7 +219,7 @@ describe('PropertyViewComponent', () => {
 
         it('should show an add button under each property that has a value component and for which the cardinality is not 1', () => {
             const addButtons = propertyViewComponentDe.queryAll(By.css('button.create'));
-            expect(addButtons.length).toEqual(14);
+            expect(addButtons.length).toEqual(13);
 
         });
 
@@ -231,7 +236,7 @@ describe('PropertyViewComponent', () => {
             const addButtons = propertyViewComponentDe.queryAll(By.css('button.create'));
 
             // the add button for the property with the open add value form is hidden
-            expect(addButtons.length).toEqual(13);
+            expect(addButtons.length).toEqual(12);
 
             expect(propertyViewComponentDe.query(By.css('.add-value'))).toBeDefined();