+++ src/app/action-playground/action-playground.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { ApiResponseError, StringLiteral } from '@dasch-swiss/dsp-js';
-import { DspMessageData, SortingService } from '@dasch-swiss/dsp-ui';
-import { MatSliderChange } from '@angular/material/slider';
+import { ConfirmationDialogComponent, ConfirmationDialogData, DspMessageData, SortingService } from '@dasch-swiss/dsp-ui';
 
 @Component({
   selector: 'app-action-playground',
@@ -133,18 +133,16 @@ export class ActionPlaygroundComponent implements OnInit {
 
     // time for timecode pipe
     time = 342;
+    confirmationDialogResponse: string;
+    showTimedMessage: boolean;
 
     constructor(
-        // private _sessionService: SessionService,
-        // @Inject(DspApiConnectionToken) private dspApiConnection: KnoraApiConnection,
-        private _sortingService: SortingService
+        private _sortingService: SortingService,
+        private _dialog: MatDialog
     ) { }
 
     ngOnInit(): void {
         this.refresh();
-
-        // already logged-in user?
-        // this.session = this._sessionService.getSession();
     }
 
     // only for testing the change of status
@@ -175,42 +173,29 @@ export class ActionPlaygroundComponent implements OnInit {
         console.log('submit string literal', this.stringLiteralInputNewLabels);
     }
 
-    // TODO: Will be replaced by login process from action module
-    /*
-    login() {
-        this.loading = true;
-        this.dspApiConnection.v2.auth.login('username', 'root', 'test').subscribe(
-            (response: ApiResponseData<LoginResponse>) => {
-                this._sessionService.setSession(response.body.token, 'root', 'username').subscribe(
-                    () => {
-                        this.loading = false;
-                        this.session = this._sessionService.getSession();
-                    });
-            },
-            (error: ApiResponseError) => {
-                // error handling
-                // this.loginErrorUser = (error.status === 404);
-                // this.loginErrorPw = (error.status === 401);
-                // this.loginErrorServer = (error.status === 0);
+    // confirmation dialog
 
-                // this.errorMessage = error;
+    openDialog() {
+        const dialogData = new ConfirmationDialogData();
+        dialogData.title = 'Are you sure want to do this?';
+        dialogData.message = 'Confirming this action will delete the value. (Not really though, this is just a test message)';
+        dialogData.buttonTextOk = 'Yes, delete the value';
+        dialogData.buttonTextCancel = 'No, keep the value';
 
-                this.loading = false;
-                // TODO: update error handling similar to the old method (see commented code below)
+        const dialogRef =
+            this._dialog.open<ConfirmationDialogComponent, ConfirmationDialogData>(ConfirmationDialogComponent, { data: dialogData});
+
+        dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+            if (confirmed) {
+                this.confirmationDialogResponse = 'Action was confirmed!';
+            } else {
+                this.confirmationDialogResponse = 'Action was not confirmed';
             }
-        );
+        });
     }
 
-    // TODO: Will be replaced by login process from action module
-    logout() {
-        this.loading = true;
-        this.dspApiConnection.v2.auth.logout().subscribe(
-            (response: ApiResponseData<LogoutResponse>) => {
-                this._sessionService.destroySession();
-                this.session = this._sessionService.getSession();
-                this.loading = false;
-            }
-        )
+    openMessage() {
+        this.showTimedMessage = true;
+        setTimeout(() => { this.showTimedMessage = false; }, 2100);
     }
-    */
 }